# Tests: reuse network namespaces