## Changes

Reuse network namespaces during integration tests. This avoids spending time creating network devices and namespaces, which is a slow operation because of networking lock contention.

After doing that, since we don't have lock contention anymore, we can double the number of test workers and see an ~25% improvement on top of that.

## Reason

To make functional test runs faster. Though this is low priority and the functional tests are not the long pole, this change is relatively small and self-contained.

The 2 changes together almost half the execution time:

| instance | before (s) | after (s) | % |
|:-:|-:|-:|-:|
| m6a | 361 | 195 | -47% |
|m6g | 245 | 137 |-45% |

## License Acceptance

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license. For more information on following Developer
Certificate of Origin and signing off your commits, please check
[`CONTRIBUTING.md`][3].

## PR Checklist

- [x] I have read and understand [CONTRIBUTING.md][3].
- [x] I have run `tools/devtool checkstyle` to verify that the PR passes the
  automated style checks.
- [x] I have described what is done in these changes, why they are needed, and
  how they are solving the problem in a clear and encompassing way.
- [x] I have updated any relevant documentation (both in code and in the docs)
  in the PR.
- [x] I have mentioned all user-facing changes in `CHANGELOG.md`.
- [x] If a specific issue led to this PR, this PR closes the issue.
- [x] When making API changes, I have followed the
  [Runbook for Firecracker API changes][2].
- [x] I have tested all new and changed functionalities in unit tests and/or
  integration tests.
- [x] I have linked an issue to every new `TODO`.

______________________________________________________________________

- [x] This functionality cannot be added in [`rust-vmm`][1].

[1]: https://github.com/rust-vmm
[2]: https://github.com/firecracker-microvm/firecracker/blob/main/docs/api-change-runbook.md
[3]: https://github.com/firecracker-microvm/firecracker/blob/main/CONTRIBUTING.md